+++ Network/NW.go~
@@ -1,7 +1,14 @@
 package main
 
 
-
+import (
+	"./localip"
+	"./peers"
+	"flag"
+	"fmt"
+	"os"
+	"time"
+)
